pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "document-intelligence-refinery"
version = "0.1.0"
description = "A production-grade, multi-stage agentic pipeline for unstructured document extraction"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "pdfplumber>=0.10.0",
    "pydantic>=2.5.0",
    "PyYAML>=6.0",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
layout = [
    "docling>=2.0.0",
]
vision = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
